moved LineageContext struct to domain/user
