+++ crates/api_models/src/user.rs
@@ -396,13 +396,3 @@ pub struct ListProfilesForUserInOrgAndMerchantAccountResponse {
     pub profile_id: id_type::ProfileId,
     pub profile_name: String,
 }
-
-#[derive(Debug, serde::Serialize, serde::Deserialize)]
-pub struct LineageContext {
-    pub user_id: String,
-    pub merchant_id: id_type::MerchantId,
-    pub role_id: String,
-    pub org_id: id_type::OrganizationId,
-    pub profile_id: id_type::ProfileId,
-    pub tenant_id: Option<id_type::TenantId>,
-}
